% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.r
\name{as_document_network}
\alias{as_document_network}
\title{Create a document similarity network}
\usage{
as_document_network(el)
}
\arguments{
\item{el}{An RNewsflow_edgelist object, as created with \link[RNewsflow]{compare_documents}.}
}
\value{
A network/graph in the \link[igraph]{igraph} class
}
\description{
This function can be used to structure the output of the \link[RNewsflow]{compare_documents} function as an igraph network.
}
\examples{
rnewsflow_dfm 

dtm = quanteda::dfm_tfidf(rnewsflow_dfm)
el = compare_documents(dtm, date_var='date', hour_window = c(0.1, 36))

g = as_document_network(el)
g
}
