% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.r
\name{directed.network.plot}
\alias{directed.network.plot}
\title{A wrapper for \link[igraph]{plot.igraph} for visualizing directed networks.}
\usage{
directed.network.plot(g, weight.var = "from.Vprop", weight.thres = NULL,
  delete.isolates = F, vertex.size = 30, vertex.color = "lightblue",
  vertex.label.color = "black", vertex.label.cex = 0.7,
  edge.color = "grey", show.edge.labels = T, edge.label.color = "black",
  edge.label.cex = 0.6, edge.arrow.size = 1,
  layout = layout.davidson.harel, ...)
}
\arguments{
\item{g}{A network/graph in the \link[igraph]{igraph} class}

\item{weight.var}{The edge attribute that is used to specify the edges}

\item{weight.thres}{A threshold for weight. Edges below the threshold are ignored}

\item{delete.isolates}{If TRUE, isolates (i.e. vertices without edges) are ignored.}

\item{...}{Arguments to be passed to the \link[igraph]{plot.igraph} function.}
}
\value{
Nothing
}
\description{
This is a convenience function for visualizing directed networks with edge labels using \link[igraph]{plot.igraph}. 
It was designed specifically for visualizing aggregated document similarity networks in the \link[rNewsflow]{rNewsflow} package, but works with any network in the \link[igraph]{igraph} class.
}
\examples{
data(docnet)
aggdocnet = network.aggregate(docnet, by='source')
directed.network.plot(aggdocnet, weight.var = 'to.Vprop', weight.thres = 0.2)
}

