% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{docnet}
\alias{docnet}
\title{Document similarity network for one news agency, and the print and online editions of two newspapers}
\format{
docnet: A network/graph in the \link[igraph]{igraph} class as created with \link[RNewsflow]{document.network} or \link[RNewsflow]{newsflow.compare}.
}
\description{
Document similarity network for one news agency, and the print and online editions of two newspapers
}
\keyword{datasets}
