% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compare.r
\name{documents.compare}
\alias{documents.compare}
\title{Compare the documents in two corpora/dtms}
\usage{
documents.compare(dtm, dtm.y = NULL, measure = "cosine",
  min.similarity = 0.1, n.topsim = NULL, return.zeros = F)
}
\arguments{
\item{dtm}{a document-term matrix in format of the tm package.}

\item{dtm.y}{Optional. If given, documents from dtm will only be compared to the documents in dtm.y}

\item{measure}{the measure that should be used to calculate similarity/distance/adjacency. Currently only cosine is supported}

\item{min.similarity}{a threshold for similarity. lower values are deleted. Set to 0.1 by default.}

\item{n.topsim}{An alternative or additional sort of threshold for similarity. Only keep the [n.topsim] highest similarity scores for x. Can return more than [n.topsim] similarity scores in the case of duplicate similarities.}

\item{return.zeros}{If true, all comparison results are returned, including those with zero similarity (quite possibly the worst thing to do with large data)}
}
\value{
A data frame with sets of documents and their similarities.
}
\description{
Compare the documents in corpus dtm.x with reference corpus dtm.y.
}

