% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compare.r
\name{documents.window.compare}
\alias{documents.window.compare}
\title{Compare the documents in a dtm per time frame}
\usage{
documents.window.compare(dtm, meta, id.var = "document_id",
  date.var = "date", hour.window = c(-24, 24), measure = "cosine",
  min.similarity = 0, n.topsim = NULL, only.from = NULL, only.to = NULL,
  return.zeros = F, only.complete.window = T)
}
\arguments{
\item{dtm}{a document-term matrix in format of the tm package.}

\item{measure}{the measure that should be used to calculate similarity/distance/adjacency. Currently only cosine is supported}

\item{min.similarity}{a threshold for similarity. lower values are deleted. Set to 0.1 by default.}

\item{n.topsim}{An alternative or additional sort of threshold for similarity. Only keep the [n.topsim] highest similarity scores for x. Can return more than [n.topsim] similarity scores in the case of duplicate similarities.}

\item{only.from}{A vector with names/ids of documents (dtm rownames), or a logical vector that matches the rows of the dtm. Use to compare only these documents to other documents.}

\item{only.to}{A vector with names/ids of documents (dtm rownames), or a logical vector that matches the rows of the dtm. Use to compare other documents to only these documents.}

\item{return.zeros}{If true, all comparison results are returned, including those with zero similarity (quite possibly the worst thing to do with large data)}

\item{only.complete.window}{if True, only compare articles (x) of which a full window of reference articles (y) is available. Thus, for the first and last [window.size] days, there will be no results for x.}

\item{date}{a vector of date class, of the same length and order as the documents (rows) of the dtm.}

\item{window.size}{the timeframe in days within which articles must occur in order to be compared. e.g., if 0, articles are only compared to articles of the same day. If 1, articles are compared to all articles of the previous, same or next day.}

\item{window.direction}{For a more specific selection of which articles in the window to compare to. This is given with a combination of the symbols '<' (before x) '=' (simultanous with x) and '>' (after x). default is '<=>', which means all articles. '<>' means all articles before or after the [time.unit] of an article itself. '<' means all previous articles, and '<=' means all previous and simultaneous articles. etc.}

\item{return.date}{If true, the dates for x and y are given in the output}
}
\value{
A data frame with columns x, y and similarity. If return.date == T, date.x and date.y are returned as well.
}
\description{
This is an extension of the documents.compare function to only compare documents within a given range of days.
}
\details{
Note that in this function it is not possible to use dtm.y to compare only certain documents (e.g., news and PR messages). Instead, the only.from and only.to parameters can be used for this purpose.
}

