% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{dtm}
\alias{dtm}
\title{Document Term Matrix for one news agency, and the print and online editions of two newspapers}
\format{dtm: A document term matrix in the `DocumentTermMatrix` class of the tm package.}
\usage{
data(dtm)
}
\description{
Document Term Matrix for one news agency, and the print and online editions of two newspapers
}
\keyword{datasets}

