% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_preparation.r
\name{get_overlap_terms}
\alias{get_overlap_terms}
\title{View overlapping terms for a given pair of documents}
\usage{
get_overlap_terms(dtm, doc.x, doc.y, dtm.y = dtm)
}
\arguments{
\item{dtm}{A quanteda dfm}

\item{doc.x}{The name of the first document in dtm}

\item{doc.y}{The name of the second document in dtm (or dtm.y)}

\item{dtm.y}{Optionally, a second dtm (for when the documents occur in separate dtm's)}
}
\value{
A character vector
}
\description{
View overlapping terms for a given pair of documents
}
\examples{
get_overlap_terms(rnewsflow_dfm, 
                  quanteda::docnames(rnewsflow_dfm)[1],
                  quanteda::docnames(rnewsflow_dfm)[5])
}
