% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{meta}
\alias{meta}
\title{Meta information for one news agency, and the print and online editions of two newspapers}
\format{meta: A data.frame containing the meta information for each document.}
\usage{
data(meta)
}
\description{
Meta information for one news agency, and the print and online editions of two newspapers
}
\keyword{datasets}

