% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{rnewsflow_dfm}
\alias{rnewsflow_dfm}
\title{quanteda dfm for RNewsflow vignette demo}
\format{dfm}
\usage{
rnewsflow_dfm
}
\description{
quanteda dfm for RNewsflow vignette demo
}
\keyword{datasets}
