% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{terms}
\alias{terms}
\title{Preprocessed terms per document for one news agency, and the print and online editions of two newspapers}
\format{terms: A data.frame containing preprocessed terms per document.}
\usage{
data(terms)
}
\description{
Preprocessed terms per document for one news agency, and the print and online editions of two newspapers
}
\keyword{datasets}

