% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_preparation.r
\name{create_queries}
\alias{create_queries}
\title{(experimental) Combine column in a document term matrix to get highly sparse, dichotomous features.}
\usage{
create_queries(dtm, ref_dtm = NULL, min_docfreq = 2,
  max_docprob = 0.001, weight = T, union_sim_thres = NA,
  verbose = T, only_dtm_combs = T)
}
\arguments{
\item{dtm}{A quanteda \link[quanteda]{dfm}}

\item{ref_dtm}{Optionally, another quanteda \link[quanteda]{dfm}. If given, the ref_dtm will be used to calculate the docfreq/docprob scores.}

\item{min_docfreq}{The minimum frequency for terms or combinations of terms}

\item{max_docprob}{The maximum probability (document frequency / N) for terms or combinations of terms}

\item{weight}{If true, weight the query dtm (if ref_dtm is used, uses the idf of the ref_dtm)}

\item{union_sim_thres}{If given, a number between 0 and 1, used as the cosine similarity threshold for combining clusters of terms}

\item{verbose}{If true, report progress}

\item{only_dtm_combs}{Only include term combinations that occur in dtm}
}
\value{
a list with a query dtm and lookup dtm.
}
\description{
Creates queries
}
\details{
Performs two operations. 
First, clusters of very similar columns (high cosine similarity) can be merged into a single column.
This is an OR (union) combination, meaning that if at least one column is nonzero, the value will be one.
Second, all columns will be combined to get the co-occurences (AND, or intersect).

The output dtm will have dichotomous features (the cluster of combination of clusters does or does not occur in the document).
To keep the size manageable, a max_docprob is used to only keep features that occur in at most max_docprob*nrow(dtm) documents.

If a ref_dtm is given, only the ref_dtm will be used to compute the new features.
The output will contain both the dtm and ref_dtm transformed to contain only the new features
}
\examples{
 q = create_queries(rnewsflow_dfm, min_docfreq = 2, union_sim_thres = 0.9, 
                    max_docprob = 0.05, verbose = FALSE)
 head(colnames(q$query_dtm),100)
}
