% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compare.r
\name{delete.duplicates}
\alias{delete.duplicates}
\title{Delete duplicate (or similar) documents from a document term matrix}
\usage{
delete.duplicates(dtm, meta, id.var = "document_id", date.var = "date",
  source.var = "source", hour.window = c(-24, 24), similarity = 1,
  keep = "first", tf.idf = F)
}
\arguments{
\item{dtm}{a document-term matrix in format of the tm package.}

\item{similarity}{a threshold for similarity. lower values are deleted}

\item{keep}{A character indicating whether to keep the 'first' or 'last' of duplicate documents.}

\item{tf.idf}{if True, weight the dtm with tf.idf before comparing documents}

\item{date}{a vector of date class, of the same length and order as the documents (rows) of the dtm.}

\item{source}{a vector of the same length and order as the documents (rows) of the dtm, indicating what the source of the document is.}

\item{window.size}{the timeframe in days within which articles must occur in order to be compared. e.g., if 0, articles are only compared to articles of the same day. If 1, articles are compared to all articles of the previous, same or next day.}
}
\value{
A dtm with the duplicate documents deleted
}
\description{
Delete duplicate (or similar) documents from a document term matrix
}

