% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_preparation.r
\name{term_intersect}
\alias{term_intersect}
\title{Combine terms in a dtm}
\usage{
term_intersect(dtm, simmat, as_dfm = T, verbose = F)
}
\arguments{
\item{dtm}{A quanteda \link[quanteda]{dfm} or a dgCMatrix.}

\item{simmat}{A similarity matrix in dgCMatrix format. For instance, created with \link{term_char_sim}}

\item{as_dfm}{If True, return as quanteda dfm}

\item{verbose}{If True, report progress}
}
\value{
A dgCMatrix or quanteda dfm
}
\description{
Given a dtm and a similarity (adjacency) matrix, create a new column for each nonzero cell in the
similarity matrix. For the term combinations  (everything except the diagonal) the column names will be
pasted together with a "&" separator (read as AND)
}
